Breaking change: change field names
Change the field name of "git" to "receiving_url" and "gitdir" to
"install_dir". This is to make the fields consistent among different
installation methods (git, url and tar) and to make their names more
descriptive.
Since the project has not yet been announced, this is about the last
possible time to make a big breaking change like this.